images: Ensure that the useradd(8), etc. manuals are available
The shadow-utils package has always been part of the fedora base image.
It's explicitly listed in extra-packages as a safeguard against losing
useradd(8) and usermod(8) by mistake because they are needed by the
entry point of a Toolbx container [1].  Hence, the need to restore the
shadow-utils documentation that was stripped out in the base image.

Only the images for currently maintained Fedoras (ie., 37, 38 and 39)
were updated.

[1] Commit c6772f0f112e8004
    https://github.com/containers/toolbox/commit/c6772f0f112e8004

https://github.com/containers/toolbox/pull/1394