cmd/create, pkg/skopeo: Make the 'skopeo inspect' invocation cancellable
A subsequent commit will use this to ensure that the user can still
interact with the image download prompt while 'skopeo inspect' fetches
the image size from the remote registry.  Initially, the prompt will be
shown without the image size.  If the user responds before the size is
fetched, then the pending 'skopeo inspect' will be cancelled.

https://github.com/containers/toolbox/issues/752
https://github.com/containers/toolbox/issues/1263