pkg/term: Add ways to change the state of a terminal device
A subsequent commit will use this to ensure that the user can still
interact with the image download prompt while 'skopeo inspect' fetches
the image size from the remote registry.  To do this, at some point, the
terminal device will be put into non-canonical mode input and the
echoing of input characters will be disabled to retain full control of
the cursor position.

https://github.com/containers/toolbox/issues/752
https://github.com/containers/toolbox/issues/1263