cmd/initContainer: Defend against insufficient resources for inotify(7)
Currently, inotify(7) is used to keep /etc/timezone inside the Toolbx
container synchronized with the host operating system's /etc/localtime.
However, /etc/timezone is only there for compatibility with Java.  The
vast majority of non-ancient code bases use /etc/localtime, which does
not need inotify(7) to stay synchronized.

Therefore, it's not worth preventing the container from starting when
the operating system is suffering from a shortage of resources needed
for inotify(7).  Especially because this shortage can be caused by a bug
in another program that's consuming too many inotify(7) instances and
watches.

https://github.com/containers/toolbox/issues/1329